basket error

+++ telegram/handlers.py
@@ -60,8 +60,9 @@ def handle_callback_query(cq: types.CallbackQuery):
 
     old_state = control.user_session.state
     get_handler(control).handle(cq.data, cq.message.id)
-    bot.delete_message(cq.message.chat.id, cq.message.id)
-    get_displayer(control).show()
+    if old_state != control.user_session.state:
+        bot.delete_message(cq.message.chat.id, cq.message.id)
+        get_displayer(control).show()
 
 
 @bot.message_handler(content_types=['location'])
